Make models closable
The design now given the model ownership of the HDF5 file and
responsibility for closing it. This will allow more lazy loading in
future. The fetcher in turn is responsible for closing the model.

The fetch_models function wasn't easily to use correctly with closable
models, so has been removed.