WIP: extract a base class out of Fetcher
I still plan to reorganise the package structure a bit later; this is
proof-of-concept work and needs documentation and some re-thinking. The
idea is that the base class does the work in coroutines, yielding
objects to represent simplified HTTP requests and having responds
returned to it by the caller. That makes it much easier to put both
synchronous and asynchronous wrappers around it.