Uniform support for file URLs
Instead of using requests_file (which doesn't work with aiohttp), do a
home-grown version. This also allows h5py to be given the on-disk file
when using a file:// URL, rather than giving it a BytesIO.

A side effect of this is that an external alias file can't do an HTTP
redirect to a local file. That's probably a good thing for security.