Do automatic HTTP retries
This is done when the session is created by default. Retries are tried
up to 5 total attempts (including the initial one) on connection errors
and HTTP status 5xx. Unfortunately the logic is slightly different
between the requests and aiohttp backends because aiohttp doesn't have
native retry support and hence it is provided by aiohttp-retry, which
requires the retry options to be passed to each request rather than
stored in the class.