Improve documentation for RFIMask
The interaction of `mask_auto_correlations` and `max_baseline_length`
in particular did not quite make sense to me, so I fleshed out their
descriptions a bit more. Point out that the `max_baseline` column in
the table should have non-negative values (otherwise the corresponding
ranges will be effectively disabled and might as well be removed).