Make beam sampling work with multi-dimensional l/m
Another major change is to the axis order in the file. Frequency is
moved to the front. This was done mainly so that the result of
scipy.interpolate.interp1d would be C-contiguous (making numba's @ more
efficient), but it also ensures that the data used at a single frequency
is contiguous, potentially simplifying lazy loading in future.