Refactoring to support vectorised RADecFrame
In theory it should now all work, but there are no tests. To simplify
things, the application of Jones matrices to rotate polarization from HV
to XY is now done after the Fourier transform. When sampling large
numbers of coordinates in a single frame this could be more expensive,
but is still linear-cost in the size of the output so should not
contribute significantly.

Implements SPR1-969.